♻️ refactor: 리뷰 반영 - 사이드바 컴포넌트 분리(thanks @elbica) [YW-357]
